requirements.txt
anthropic>=0.87.0
datasets>=4.8.0
matplotlib>=3.9.0
numpy>=2.0.0
openai>=2.0.0
python-dotenv>=1.0.0
